pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "worldgen"
version = "0.1.0"
authors = [
  { name="Ziyang Xie", email="ziyangxie01@gmail.com" },
]
description = "A package for 3D scene generation 🌏"
readme = "README.md"
requires-python = ">=3.11" 
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "diffusers>=0.33.1",
    "xformers>=0.0.29.post3",
    "transformers>=4.48.3",
    "viser==0.2.23",
    "py360convert>=0.1.0",
    "einops>=0.7.0",
    "pillow>=8.0.0",
    "UniK3D @ git+https://github.com/lpiccinelli-eth/UniK3D.git",
    "pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git",
    "scikit-image>=0.24.0",
    "sentencepiece>=0.2.0",
    "peft>=0.7.1"
]

[tool.setuptools.packages.find]
where = ["src"] 

[tool.setuptools.dynamic]
version = {attr = "worldgen.__version__"} 